Release 4.1.8
* Update README and CommonUtilities submodule

Enhanced the README with improved structure, updated project description, features, installation, and usage instructions. Changed license references to MIT, added sponsorship and acknowledgments, and updated the CommonUtilities submodule to the latest commit.

* Bump version to 4.1.8

Updated AssemblyVersion and FileVersion in EasyKit.csproj and installer scripts for both x64 and x86 to 4.1.8 to prepare for a new release.